requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
attrs==25.3.0
beautifulsoup4==4.14.2
beir==2.2.0
bitarray==3.4.2
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
codecarbon==3.0.2
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.3
datasets==3.6.0
dill==0.3.8
faiss-cpu==1.11.0
fief-client==0.20.0
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.0
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.32.4
humanfriendly==10.0
idna==3.10
inquirerpy==0.3.4
isort==6.0.1
Jinja2==3.1.6
joblib==1.5.1
jwcrypto==1.5.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.5
ninja==1.11.1.4
nltk==3.9.2
numpy==2.2.6
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
ollama==0.6.0
onnxruntime==1.22.0
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pfzy==0.3.4
pillow==11.2.1
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyjnius==1.6.1
pynvml==12.0.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytrec_eval-terrier==0.5.7
pytz==2025.2
PyYAML==6.0.2
questionary==2.1.0
rank-bm25==0.2.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sympy==1.14.0
tabulate==0.9.0
termcolor==2.3.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0+cu128
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
wcwidth==0.2.13
wikiextractor==3.0.6
xxhash==3.5.0
yarl==1.20.0
yaspin==3.1.0
